+++ repos/core/src/pages/eis/utils.js
@@ -1,5 +1,6 @@
 import { merge, isObject } from 'ui-utils-pack/object'
 import { errorsMap } from './rules'
+import { FIELD } from 'ui-modules-pack'
 
 export const getFormsData = (forms) => {
   const formDataArray = [];
@@ -22,7 +23,7 @@ const getStructuredDataFromFormObject = (form, meta) => {
     return;
   }
 
-  const formState = form.getState().values;
+  const formState = changeOptionOrderForSelectFields(form.getState().values, meta);
 
   if (!relativePath) {
     if (formState.dataKind) {
@@ -183,4 +184,39 @@ export const convertFieldNameToTitleCaseText = (str) => {
   const result = fieldName.replace(/([A-Z])/g, " $1").trim();
 
   return result.charAt(0).toUpperCase() + result.slice(1);
+}
+
+// Find Select fields and change options order in case select was changed
+const changeOptionOrderForSelectFields = (data, meta) => {
+  // find data related to Select and change options order
+  const recursiveDataParser = (data2, optionName, selectValue) => {
+    let isDataOrderChanged = false
+    if (isObject(data2)) {
+      Object.keys(data2).forEach(key => {
+        if (Array.isArray(data2[key]) && data2[key][0] && data2[key][0][optionName]) {
+          data2[key].unshift(data2[key].splice(selectValue, 1)[0]);
+          isDataOrderChanged = true
+        }
+      })
+    }
+    return isDataOrderChanged
+  }
+
+  if (meta.view === FIELD.TYPE.SELECT) {
+    const selectName = meta.name;
+    if (typeof data[selectName] === 'string') {
+      const { mapOptions: { text: optionName, value: optionValue } } = meta;
+      if (optionValue === '{index}') {
+        recursiveDataParser(data, optionName, data[selectName]) && delete data[selectName]
+      }
+    }
+  }
+
+  if (Array.isArray(meta.items)) {
+    meta.items.forEach(item => {
+      changeOptionOrderForSelectFields(data, item)
+    })
+  }
+
+  return data;
 }